pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "integrator"
version = "0.1.0"
description = "A package for integrating X-ray diffraction data with variational inference."
readme = "README.md"
requires-python = ">=3.12"
authors = [
  { name = "Luis Aldama", email = "luis_aldama@g.harvard.edu" },
  { name = "Kevin Dalton", email = "kmdalton@slac.stanford.edu" },
]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]

dependencies = [
  "matplotlib>=3.5",
  "numpy>=2.3",
  "pandas>=2.3.2",
  "polars>=1.6.0",
  "pytorch_lightning>=2.2",
  "reciprocalspaceship>=1.0.7",
  "rs_distributions>=0.0.3",
  "scipy>=1.11",
]

[project.optional-dependencies]
dev = ['pytest','ruff','mypy','wandb>=0.17']

[project.scripts]
"integrator.train" = "integrator.cli.train:main"
"integrator.pred" = "integrator.cli.pred:main"

[project.urls]
Homepage = "https://github.com/Hekstra-Lab/integrator/tree/integrator_mvn"
Repository = "https://github.com/Hekstra-Lab/integrator"

# ----------------------------
# Hatchling build config
# ----------------------------
# Your code uses a "src" layout. Point hatchling at the package folder.
[tool.hatch.build.targets.wheel]
packages = ["src/integrator"]

#include = []

# Source distribution: include your source tree and top-level files.
[tool.hatch.build.targets.sdist]
include = [
  "src/**",
  "docs/**",
  "tests/**",        # optional—include tests in the sdist if you want
  "README.md",
  "pyproject.toml",
  "mkdocs.yml"
]

# ----------------------------
# Tool configs (same as before)
# ----------------------------

[tool.hatch.version]

[tool.ruff]
target-version = "py312"
line-length = 79
extend-select = ["B","I","ARG","UP","F"]
extend-exclude = [
  "build",
  "dist",
  "scratch",
  ".venv",
  "**/__pycache__",
  ".mypy_cache",
  ".ruff_cache",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true


[tool.ruff.lint]
# Only ignore variables named "_".
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.mypy]
python_version = "3.12"
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = ["tests"]

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" }
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true


[dependency-groups]
docs = [
  "mkdocs>=1.6",
  "mkdocs-material>=9.5",
  "mkdocstrings[python]>=0.18",
  "mkdocs-mermaid2-plugin>=1.1",
]
dev = [
    "griffe-inherited-docstrings>=1.1.1",
    "ipython>=8.37.0",
    "mypy>=1.17.1",
    "plotly>=6.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.9",
    "wandb>=0.23.1",
]
